requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
pinecone==3.0.0
python-dotenv==1.0.0
pydantic==2.5.0
torch==2.1.0
transformers==4.31.0
huggingface-hub==0.16.4
numpy==1.24.3
safetensors==0.4.0